+++ src/main/java/fr/upem/aquarium/Aquarium/ressource/AnimalRessource.java
@@ -5,7 +5,10 @@ import fr.upem.aquarium.Aquarium.model.Espece;
 import fr.upem.aquarium.Aquarium.service.AnimalService;
 import fr.upem.aquarium.Aquarium.service.EspeceService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -22,18 +25,29 @@ public class AnimalRessource {
     }
 
     @PostMapping("/animal")
-    public Animal postAnimal(@RequestBody Animal animal) {
-        return animalService.createAnimal(animal);
+    public ResponseEntity<Animal> postAnimal(@RequestBody Animal animal) {
+
+        return new ResponseEntity<>(animalService.createAnimal(animal), HttpStatus.CREATED);
+
     }
 
     @GetMapping("animal/{id}")
     public Optional<Animal> getOne(@PathVariable Long id) {
         //@PathVariable {id}
+        if(!animalService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Animal avec l'id " + id + " n'existe pas");
+
+
+
         return animalService.getOne(id);
     }
 
     @GetMapping("/animauxOfEspece/{idEspece}")
     public Iterable<Animal> getAnimauxOfEspece(@PathVariable Long idEspece) {
+
+        if(!especeService.getOne(idEspece).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Espèce avec l'id " + idEspece + " n'existe pas");
+
 
 
         return animalService.getAnimauxOfEspece(idEspece);
@@ -41,14 +55,25 @@ public class AnimalRessource {
 
 
     @PostMapping("/animal_espece/{id}")
-    public Animal postAnimalWithEspece(@RequestBody Animal animal, @PathVariable Long id) {
+    public ResponseEntity<Animal> postAnimalWithEspece(@RequestBody Animal animal, @PathVariable Long id) {
+        if(!especeService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Espèce avec l'id " + id + " n'existe pas");
+
+
+
         Optional<Espece> e = (especeService.getOne(id));
-        return animalService.createAnimalEspece(animal, e.get());
+        return new ResponseEntity<>(animalService.createAnimalEspece(animal, e.get()), HttpStatus.CREATED);
+
     }
 
 
     @DeleteMapping("animal/{id}")
     public void deleteAnimal(@PathVariable Long id) {
+
+        if(!animalService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Animal avec l'id " + id + " n'existe pas");
+
+
         animalService.deleteAnimal(id);
     }
 
@@ -57,6 +82,11 @@ public class AnimalRessource {
 
     @PostMapping("animal/{id}")
     public Animal putAnimal(@PathVariable Long id, @RequestBody Animal animal) {
+        if(!animalService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Animal avec l'id " + id + " n'existe pas");
+
+
+
         return animalService.updateAnimal(id, animal);
     }
 }
