+++ src/main/java/fr/upem/aquarium/Aquarium/ressource/EmployeRessource.java
@@ -9,6 +9,7 @@ import fr.upem.aquarium.Aquarium.service.EspeceService;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -25,18 +26,29 @@ public class EmployeRessource {
     }
 
     @PostMapping("/employe")
-    public Employe postEmploye(@RequestBody Employe employe) {
-        return employeService.createEmploye(employe);
+    public ResponseEntity<Employe> postEmploye(@RequestBody Employe employe) {
+        return new ResponseEntity<>(employeService.createEmploye(employe), HttpStatus.CREATED);
+
     }
 
     @GetMapping("employe/{id}")
     public Optional<Employe> getOne(@PathVariable Long id) {
         //@PathVariable {id}
+        if(!employeService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employé avec l'id " + id + " n'existe pas");
+
         return employeService.getOne(id);
+
+
+
     }
 
     @DeleteMapping("employe/{id}")
     public void deleteEmploye(@PathVariable Long id) {
+        if(!employeService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employé avec l'id " + id + " n'existe pas");
+
+
         employeService.deleteEmploye(id);
     }
 
@@ -45,6 +57,9 @@ public class EmployeRessource {
 
     @PostMapping("employe/{id}")
     public Employe putEmploye(@PathVariable Long id, @RequestBody Employe employe) {
+        if(!employeService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employé avec l'id " + id + " n'existe pas");
+
         return employeService.updateEmploye(id, employe);
     }
