+++ src/main/java/fr/upem/aquarium/Aquarium/ressource/SecteurRessource.java
@@ -9,7 +9,10 @@ import fr.upem.aquarium.Aquarium.service.BassinService;
 import fr.upem.aquarium.Aquarium.service.EspeceService;
 import fr.upem.aquarium.Aquarium.service.SecteurService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -26,18 +29,27 @@ public class SecteurRessource {
     }
 
     @PostMapping("/secteur")
-    public Secteur postSecteur(@RequestBody Secteur secteur) {
-        return secteurService.createSecteur(secteur);
+    public ResponseEntity<Secteur> postSecteur(@RequestBody Secteur secteur) {
+
+        return new ResponseEntity<>(secteurService.createSecteur(secteur), HttpStatus.CREATED);
+
     }
 
     @GetMapping("secteur/{id}")
     public Optional<Secteur> getOne(@PathVariable Long id) {
         //@PathVariable {id}
+        if(!secteurService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Secteur avec l'id " + id + " n'existe pas");
+
         return secteurService.getOne(id);
     }
 
     @GetMapping("secteurs/{id}/{id2}")
     public void assignBassinSecteur(@PathVariable Long id, @PathVariable Long id2) {
+        if(!bassinService.getOne(id2).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bassin avec l'id " + id2 + " n'existe pas");
+        if(!secteurService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Secteur avec l'id " + id + " n'existe pas");
 
         secteurService.addBassin(secteurService.getOne(id), bassinService.getOne(id2));
 
@@ -46,11 +58,20 @@ public class SecteurRessource {
 
     @DeleteMapping("secteurs/{id}/{id2}")
     public void removeBassinSecteur(@PathVariable Long id, @PathVariable Long id2){
+        if(!bassinService.getOne(id2).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bassin avec l'id " + id2 + " n'existe pas");
+        if(!secteurService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Secteur avec l'id " + id + " n'existe pas");
+
         secteurService.removeBassin(secteurService.getOne(id), bassinService.getOne(id2));
     }
 
     @DeleteMapping("secteur/{id}")
     public void deleteSecteur(@PathVariable Long id) {
+
+        if(!secteurService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Secteur avec l'id " + id + " n'existe pas");
+
         secteurService.deleteSecteur(id);
     }
 
@@ -59,12 +80,18 @@ public class SecteurRessource {
 
     @PostMapping("secteur/{id}")
     public Secteur putSecteur(@PathVariable Long id, @RequestBody Secteur secteur) {
+        if(!secteurService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Secteur avec l'id " + id + " n'existe pas");
+
         return secteurService.updateSecteur(id, secteur);
     }
 
     @GetMapping("secteurFromBassin/{idBassin}")
     public Optional<Secteur> getFromBassin(@PathVariable Long idBassin) {
         //@PathVariable {id}
+        if(!bassinService.getOne(idBassin).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Bassin avec l'id " + idBassin + " n'existe pas");
+
         return secteurService.getFromBassin(idBassin);
     }
 }
